FIX: Split the color picker in two
